pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "edusched"
version = "0.1.0"
description = "智能教育调度平台 - 为学校生成可行且优化的课程表"
authors = [
    {name = "Edusched Team", email = "team@edusched.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Education",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.35.0",
    "pydantic>=2.11.7",
    "sqlalchemy>=2.0.43",
    "alembic>=1.13.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.9",
    "redis>=6.4.0",
    "rq>=2.5.0",
    "httpx>=0.28.1",
    "ortools>=9.14.6206",
    "opentelemetry-api>=1.36.0",
    "opentelemetry-sdk>=1.36.0",
    "opentelemetry-instrumentation-fastapi>=0.44b0",
    "prometheus-client>=0.22.1",
    "sentry-sdk[fastapi]>=2.35.1",
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "python-dotenv>=1.0.0",
    "pydantic-settings>=2.1.0",
    "aiofiles>=23.2.1",
    "aiosmtplib>=3.0.1",
    "boto3>=1.34.0",
    "botocore>=1.34.0",
    "oss2>=2.18.0",
    "cos-python-sdk>=5.0.0",
    "azure-storage-blob>=12.19.0",
    "google-cloud-storage>=2.10.0",
    "minio>=7.2.0",
    "elasticsearch>=8.11.0",
    "loguru>=0.7.2",
    "structlog>=23.2.0",
    "msgpack>=1.0.7",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.5.0",
    "pre-commit>=3.4.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "factory-boy>=3.3.0",
    "faker>=19.0.0",
]

external = [
    "sendgrid>=6.10.0",
    "mailgun>=1.7.1",
    "pymemcache>=4.0.0",
    "aioredis>=2.0.1",
    "sentry-sdk>=2.35.1",
    "datadog>=0.47.0",
    "slack-sdk>=3.26.0",
    "requests>=2.31.0",
]

[project.urls]
Homepage = "https://github.com/edusched/edusched"
Documentation = "https://docs.edusched.com"
Repository = "https://github.com/edusched/edusched"
Issues = "https://github.com/edusched/edusched/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/edusched"]

[tool.black]
line-length = 88
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88
known_first_party = ["edusched"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "tests.*",
    "test.*",
]
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "--strict-markers",
    "--strict-config",
    "--cov=edusched",
    "--cov-report=term-missing",
    "--cov-report=html",
    "--cov-report=xml",
]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["src/edusched"]
omit = [
    "*/tests/*",
    "*/test_*",
    "*/__pycache__/*",
    "*/migrations/*",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "if self.debug:",
    "if settings.DEBUG",
    "raise AssertionError",
    "raise NotImplementedError",
    "if 0:",
    "if __name__ == .__main__.:",
    "class .*\\bProtocol\\):",
    "@(abc\\.)?abstractmethod",
]